ORC-1230: Move encryption utility functions to HadoopShimsCurrent
