+++ java/shims/src/java/org/apache/orc/impl/HadoopShimsPre2_7.java
@@ -19,16 +19,12 @@
 package org.apache.orc.impl;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.crypto.key.KeyProviderCryptoExtension;
-import org.apache.hadoop.crypto.key.KeyProviderFactory;
 import org.apache.hadoop.fs.FSDataInputStream;
-import org.apache.orc.EncryptionAlgorithm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.util.List;
 import java.util.Random;
 
 /**
@@ -62,49 +58,9 @@ public class HadoopShimsPre2_7 implements HadoopShims {
     return false;
   }
 
-  static String buildKeyVersionName(KeyMetadata key) {
-    return key.getKeyName() + "@" + key.getVersion();
-  }
-
-  static KeyProvider createKeyProvider(Configuration conf,
-                                       Random random) throws IOException {
-    List<org.apache.hadoop.crypto.key.KeyProvider> result =
-        KeyProviderFactory.getProviders(conf);
-    if (result.size() == 0) {
-      LOG.info("Can't get KeyProvider for ORC encryption from" +
-          " hadoop.security.key.provider.path.");
-      return new NullKeyProvider();
-    } else {
-      return new KeyProviderImpl(result.get(0), random);
-    }
-  }
-
-  /**
-   * Find the correct algorithm based on the key's metadata.
-   * @param meta the key's metadata
-   * @return the correct algorithm
-   */
-  static EncryptionAlgorithm findAlgorithm(KeyProviderCryptoExtension.Metadata meta) {
-    String cipher = meta.getCipher();
-    if (cipher.startsWith("AES/")) {
-      int bitLength = meta.getBitLength();
-      if (bitLength == 128) {
-        return EncryptionAlgorithm.AES_CTR_128;
-      } else {
-        if (bitLength != 256) {
-          LOG.info("ORC column encryption does not support " + bitLength +
-              " bit keys. Using 256 bits instead.");
-        }
-        return EncryptionAlgorithm.AES_CTR_256;
-      }
-    }
-    throw new IllegalArgumentException("ORC column encryption only supports" +
-        " AES and not " + cipher);
-  }
-
   @Override
   public KeyProvider getHadoopKeyProvider(Configuration conf,
                                           Random random) throws IOException {
-    return createKeyProvider(conf, random);
+    return HadoopShimsCurrent.createKeyProvider(conf, random);
   }
 }
